use right term: pace instead of speed

+++ arduino/Stepper-motors/src/movingComponents/stepperMotors.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 void setupMotors();
-void setLeftMotorSpeed(int32_t speed);
-void setRightMotorSpeed(int32_t speed);
+void setLeftMotorSpeed(double speed);
+void setRightMotorSpeed(double speed);
 
 #endif
